home session: added d/i conversion type functionality and fixed p conversion

+++ libft/ft_digcountbase.c
@@ -1,6 +1,8 @@
 #include "libft.h"
 
-int		ft_digcountbase(unsigned long n, int base)
+#include <stdio.h>
+
+int		ft_digcountbase(long long n, int base)
 {
 	int	count;
